ai_model_server/requirements.txt
fastapi
uvicorn
tensorflow
Pillow
requests
torch
torchaudio
transformers
accelerate
pydub
soundfile
pyctcdecode
onnxruntime
onnx==1.20.1
httpx
python-dotenv